--- src/pipeline.md
@@ -1,10 +1,0 @@
-- matrix scanner (fixed: 1000 Hz)
-  - scan
-  - > bitmap_channel
-- stream processor (inherit)
-  - debounce
-  - led
-  - map
-  - > keys_channel
-- hid reporter (fixed: 1000 Hz)
-  - report
