+++ src/processor/bitmap/mod.rs
@@ -1,0 +1,1 @@
+pub mod none;
