+++ src/processor/events/mod.rs
@@ -1,0 +1,3 @@
+pub mod none;
+pub mod replace;
+pub mod rgb;
