+++ src/processor/keymap.rs
@@ -1,21 +1,16 @@
 extern crate alloc;
-use alloc::{
-    string::{String, ToString},
-    vec::Vec,
-};
-use core::{
-    error, fmt,
-    ops::{Deref, DerefMut},
-    result,
-};
+use alloc::vec::Vec;
+use core::ops::{Deref, DerefMut};
 
 use crate::{
-    key::{Action, Key, LayerIndex},
+    key::{Action, LayerIndex},
     matrix::{Bitmap, Edge},
 };
 
+use super::Event;
+
 #[derive(Clone, Copy)]
-pub struct Mapping<
+pub struct KeyMap<
     const ROW_COUNT: usize,
     const COL_COUNT: usize,
     const LAYER_COUNT: usize,
@@ -23,7 +18,7 @@ pub struct Mapping<
 >(pub [[[Action<L>; COL_COUNT]; ROW_COUNT]; LAYER_COUNT]);
 
 impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L: LayerIndex>
-    Mapping<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
+    KeyMap<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
 {
     fn flatten(&mut self) {
         // TODO: flatten passthrough
@@ -31,7 +26,7 @@ impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L
 }
 
 impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L: LayerIndex> Deref
-    for Mapping<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
+    for KeyMap<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
 {
     type Target = [[[Action<L>; COL_COUNT]; ROW_COUNT]; LAYER_COUNT];
 
@@ -41,7 +36,7 @@ impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L
 }
 
 impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L: LayerIndex>
-    DerefMut for Mapping<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
+    DerefMut for KeyMap<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
 {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
@@ -49,42 +44,30 @@ impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L
 }
 
 impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L: LayerIndex>
-    Default for Mapping<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
+    Default for KeyMap<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
 {
     fn default() -> Self {
-        Mapping([[[Action::default(); COL_COUNT]; ROW_COUNT]; LAYER_COUNT])
+        KeyMap([[[Action::default(); COL_COUNT]; ROW_COUNT]; LAYER_COUNT])
     }
 }
 
-pub struct Event<L: LayerIndex> {
-    pub time_ticks: u64,
-    pub i: usize,
-    pub j: usize,
-    pub edge: Edge,
-    pub action: Action<L>,
-}
-
-pub trait BitmapProcessor<const ROW_COUNT: usize, const COL_COUNT: usize> {
-    fn process(&mut self, bitmap: &mut Bitmap<ROW_COUNT, COL_COUNT>) -> Result;
-}
-
-pub struct EventsMapper<
+pub struct KeyMapper<
     const ROW_COUNT: usize,
     const COL_COUNT: usize,
     const LAYER_COUNT: usize,
     L: LayerIndex,
 > {
     previous_bitmap: Bitmap<ROW_COUNT, COL_COUNT>,
-    mapping: Mapping<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>,
+    mapping: KeyMap<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>,
 }
 
 impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L: LayerIndex>
-    EventsMapper<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
+    KeyMapper<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
 {
-    pub fn new(mapping: Mapping<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>) -> Self {
+    pub fn new(mapping: KeyMap<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>) -> Self {
         let mut m = mapping.clone();
         m.flatten();
-        return EventsMapper {
+        return KeyMapper {
             previous_bitmap: Bitmap::default(),
             mapping: m,
         };
@@ -92,7 +75,7 @@ impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L
 }
 
 impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L: LayerIndex>
-    EventsMapper<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
+    KeyMapper<ROW_COUNT, COL_COUNT, LAYER_COUNT, L>
 {
     pub fn map(&mut self, bitmap: &Bitmap<ROW_COUNT, COL_COUNT>, events: &mut Vec<Event<L>>) {
         let mut provisional_events = Vec::<Event<L>>::with_capacity(10);
@@ -127,63 +110,5 @@ impl<const ROW_COUNT: usize, const COL_COUNT: usize, const LAYER_COUNT: usize, L
         }
         *events = provisional_events;
         self.previous_bitmap = *bitmap;
-    }
-}
-
-pub trait EventsProcessor<L: LayerIndex> {
-    fn process(&mut self, events: &mut Vec<Event<L>>) -> Result;
-}
-
-pub struct KeyReplacer {
-    from: Key,
-    to: Key,
-}
-
-#[allow(dead_code)]
-impl KeyReplacer {
-    pub fn new(from: Key, to: Key) -> Self {
-        return KeyReplacer { from, to };
-    }
-}
-
-impl<L: LayerIndex> EventsProcessor<L> for KeyReplacer {
-    fn process(&mut self, events: &mut Vec<Event<L>>) -> Result {
-        events.iter_mut().for_each(|e| match &mut e.action {
-            Action::Key(k) => {
-                if *k == self.from {
-                    *k = self.to;
-                }
-            }
-            _ => {}
-        });
-        return Ok(());
-    }
-}
-
-pub type Result = result::Result<(), Error>;
-
-#[derive(Debug)]
-pub struct Error {
-    msg: String,
-}
-
-#[allow(dead_code)]
-impl Error {
-    pub fn new(msg: &str) -> Self {
-        return Error {
-            msg: msg.to_string(),
-        };
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.msg)
-    }
-}
-
-impl error::Error for Error {
-    fn description(&self) -> &str {
-        &self.msg
     }
 }
